requirements.txt
# Telegram Bot API framework
python-telegram-bot>=22.0,<23.0

# Environment variable management
python-dotenv>=1.0.0,<2.0.0

# Auto-translation for captions/descriptions
deep-translator>=1.11.4,<2.0.0

# Emoji support for UI
emoji>=2.8.0,<3.0.0

# FFmpeg bindings for audio extraction
ffmpeg-python>=0.2.0,<0.3.0

# Async file I/O support
aiofiles>=23.0.0,<25.0.0

# Video downloader (pinned stable version)
yt-dlp>=2024.0.0

# JavaScript runtime for yt-dlp
yt-dlp-ejs>=0.8.0,<1.0.0

# Testing
pytest>=7.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
